marktex: massive overhaul for preamle sections
The handling of the preamble has undergone a massive overhaul to the
point where it is almost feature complete (still need to implement how
the preamble will be handled in groff and other output formats). marktex
is now smart enough to change the document class of your LaTeX documents
automatically and copies in the other parts of the preamble directly
before the \begin{document} command in the tex source. You can now write
in any documentclass and use any packages in LaTeX documents!